+++ sightparser/src/main/java/sugar/free/sightparser/applayer/messages/remote_control/StandardBolusMessage.java
@@ -8,7 +8,7 @@ public class StandardBolusMessage extends BolusMessage {
     private static final long serialVersionUID = 1L;
 
     @Setter
-    private float amount;
+    private double amount;
 
     @Override
     protected byte[] getData() throws Exception {
@@ -17,11 +17,11 @@ public class StandardBolusMessage extends BolusMessage {
         data.putShort((short) 0x1F00);
         data.putShort((short) 0x1F00);
         data.putShort((short) 0x0000);
-        data.putUInt16LE((int) (amount * 100F));
+        data.putUInt16LE((int) (amount * 100D));
         data.putShort((short) 0x0000);
         data.putShort((short) 0x0000);
         data.putShort((short) 0x0000);
-        data.putUInt16LE((int) (amount * 100F));
+        data.putUInt16LE((int) (amount * 100D));
         data.putShort((short) 0x0000);
         data.putShort((short) 0x0000);
         return data.getBytes();
